requirements.txt
wikipedia-api
sentence-transformers
faiss-cpu
transformers
torch
numpy
